Move scripts commons to devextreme-monorepo-tools
