+++ packages/devextreme-monorepo-tools/src/fs-utils.ts
