+++ packages/devextreme-monorepo-tools/src/index.ts
@@ -1,3 +1,4 @@
-export function foo() {
-  console.log('bar');
-}
+export * from './fs-utils';
+export * from './npm-utils';
+export * from './shell-utils';
+export * from './version-utils';
