+++ tools/scripts/common/monorepo-tools.ts
@@ -1,0 +1,1 @@
+export * from '../../../packages/devextreme-monorepo-tools/src';
