+++ tools/scripts/make-artifacts-package.ts
@@ -2,10 +2,10 @@ import sh from 'shelljs';
 import path from 'path';
 
 import { ARTIFACTS_DIR, INTERNAL_TOOLS_ARTIFACTS, JS_ARTIFACTS, TS_ARTIFACTS, CSS_ARTIFACTS } from './common/paths';
-import { ensureEmptyDir } from './common/fs-utils';
+import { ensureEmptyDir } from './common/monorepo-tools';
 
 import { version, license, author } from '../../package.json';
-import { npm } from './common/npm-utils';
+import { npm } from './common/monorepo-tools';
 
 const ARTIFACTS_PACKAGE_DIR = path.join(ARTIFACTS_DIR, 'devextreme-artifacts');
